chore: relocate package moit-api module
