+++ moit-api/src/main/kotlin/com/mashup/moit/controller/moit/dto/MoitDto.kt
@@ -1,4 +1,4 @@
-package com.mashup.moit.moit.controller.dto
+package com.mashup.moit.controller.moit.dto
 
 import com.mashup.moit.domain.attendance.Attendance
 import com.mashup.moit.domain.attendance.AttendanceStatus
